添加发射

+++ components/motor/motor_dm.h
@@ -85,7 +85,7 @@ struct MotorDmRxData
 class MotorDm
 {
 public:
-    void init(CAN_HandleTypeDef *hcan, uint8_t can_id, uint8_t master_id,
+    void init(CAN_HandleTypeDef *hcan, uint16_t can_id, uint16_t master_id,
               MotorDmControlMethod control_method = MOTOR_DM_CONTROL_METHOD_MIT,
               float p_max = 12.5f, float v_max = 30.0f, float t_max = 10.0f, float kp = 0.0f,
               float kd = 0.0f, uint8_t reverse = false);
